move getting apiExports to another file and made PR changes
