+++ @re-/tools/src/common.ts
@@ -10,7 +10,6 @@ export type UntilOptions = {
     intervalSeconds?: number
 }
 
-// dwadwadsapeeen
 export const until = async (
     condition: () => boolean,
     { timeoutSeconds = 10, intervalSeconds = 0.1 }: UntilOptions = {}
