+++ scripts/findUnused.ts
@@ -1,82 +1,33 @@
-import { relative } from "node:path"
-import { fromHere } from "@re-/node"
-import { BindingNamedNode, Node, Project, ts } from "ts-morph"
+import { join, relative } from "node:path"
+import { findPackageRoot, fromHere } from "@re-/node"
+import { BindingNamedNode, Project, SyntaxKind } from "ts-morph"
+import {
+    checkForIgnoreUnusedComment,
+    getPublicApiExports
+} from "./findUnusedHelpers.js"
 
 const project = new Project({
     tsConfigFilePath: fromHere("..", "tsconfig.references.json")
 })
 
-const unused: Record<string, string[]> = {}
-const snippetFileRegex = /snippets\/\w+/
-const exportAllRegex = /export \*/
+const unusedExports: Record<string, string[]> = {}
+const ignorePaths: RegExp[] = [new RegExp(join("docs", "snippets"))]
 const exportAllRenamedRegex = /\* as /
-const ignoreUnusedComment = "@ignore-unused"
-const apiExports: ApiExports[] = []
-const sourceFiles = project.getSourceFiles()
-type DeclarationInfo = {
-    filePath: string
-    sourceFileText: string
-    declarationNode: Node<ts.Node>
-    referencesLength: number
-    declarationName: string
-}
-type ApiExports = {
-    path: string
-    exportedDeclarations: string[]
-}
-const checkForIgnoreUnusedComment = ({
-    filePath,
-    sourceFileText,
-    declarationNode,
-    referencesLength,
-    declarationName
-}: DeclarationInfo) => {
-    for (const range of declarationNode.getLeadingCommentRanges()) {
-        if (
-            sourceFileText
-                .slice(range.getPos(), range.getEnd())
-                .includes(ignoreUnusedComment)
-        ) {
-            if (referencesLength > 1) {
-                console.log(
-                    `@ignore-unused wasn't needed for ${declarationName} - ${filePath}\n`
-                )
-            }
-            return true
-        }
-    }
-    return false
-}
 
-const indexFiles = sourceFiles.filter(
-    (file) => file.getBaseName() === "index.ts"
-)
-for (const file of indexFiles) {
-    const publicApiExports = file
-        .getExportDeclarations()
-        .filter((declaration) => !exportAllRegex.test(declaration.getText()))
-        .map((declaration) => {
-            return {
-                path: relative(".", declaration.getSourceFile().getFilePath()),
-                exportedDeclarations: declaration
-                    .getNamedExports()
-                    .map((namedExport) => namedExport.getName())
-            }
-        })
-    if (publicApiExports.length) {
-        apiExports.push(...publicApiExports)
-    }
-}
+const apiExports = getPublicApiExports(project)
+
 for (const sourceFile of project.getSourceFiles()) {
     const file = relative(".", sourceFile.getFilePath())
-    const sourceFileText = sourceFile.getFullText()
-    if (snippetFileRegex.test(file)) {
+    const text = sourceFile.getFullText()
+
+    if (ignorePaths.some((pathRegex) => pathRegex.test(file))) {
         continue
     }
-    const unusedInFile = []
+
+    const unusedExportsInFile = []
     for (const [name, declarations] of sourceFile.getExportedDeclarations()) {
         const references = declarations.flatMap((declaration) => {
-            if (declaration.getKindName() === "ExportSpecifier") {
+            if (declaration.isKind(SyntaxKind.ExportSpecifier)) {
                 return []
             }
             if (declaration.getSourceFile() !== sourceFile) {
@@ -86,8 +37,9 @@ for (const sourceFile of project.getSourceFiles()) {
                 return []
             }
             const knownExport = apiExports.find(
-                (apiExport) => apiExport.path === file
+                (apiExport) => apiExport.path === findPackageRoot(file)
             )
+
             if (knownExport) {
                 if (knownExport.exportedDeclarations.includes(name)) {
                     return []
@@ -99,52 +51,46 @@ for (const sourceFile of project.getSourceFiles()) {
 
             const sourceFileAndReferencesData = {
                 filePath: file,
-                sourceFileText,
+                text,
                 referencesLength: references.length,
                 declarationName: (
                     declaration as any as BindingNamedNode
                 ).getName()
             }
-            switch (declaration.getKindName()) {
-                case "VariableDeclaration":
-                    {
-                        const foundComment = checkForIgnoreUnusedComment({
-                            ...sourceFileAndReferencesData,
-                            declarationNode: declaration
-                                .getParentOrThrow()
-                                .getParentOrThrow()
-                        })
+            if (declaration.isKind(SyntaxKind.VariableDeclaration)) {
+                const hasIgnoreComment = checkForIgnoreUnusedComment({
+                    ...sourceFileAndReferencesData,
+                    declarationNode: declaration.getFirstAncestorByKindOrThrow(
+                        SyntaxKind.VariableStatement
+                    )
+                })
+                if (hasIgnoreComment) {
+                    return []
+                }
+            } else {
+                const hasIgnoreComment = checkForIgnoreUnusedComment({
+                    ...sourceFileAndReferencesData,
+                    declarationNode: declaration
+                })
 
-                        if (foundComment) {
-                            return []
-                        }
-                    }
-                    break
-                default: {
-                    const foundComment = checkForIgnoreUnusedComment({
-                        ...sourceFileAndReferencesData,
-                        declarationNode: declaration
-                    })
-
-                    if (foundComment) {
-                        return []
-                    }
+                if (hasIgnoreComment) {
+                    return []
                 }
             }
             return references
         })
         if (references.length === 1) {
-            unusedInFile.push(name)
+            unusedExportsInFile.push(name)
         }
     }
-    if (unusedInFile.length) {
-        unused[file] = unusedInFile
+    if (unusedExportsInFile.length) {
+        unusedExports[file] = unusedExportsInFile
     }
 }
 
-if (Object.keys(unused).length) {
+if (Object.keys(unusedExports).length) {
     console.error("The following unused exports must be removed:")
-    for (const [file, unusedNames] of Object.entries(unused)) {
+    for (const [file, unusedNames] of Object.entries(unusedExports)) {
         console.group(`\n${file}:`)
         for (const unusedName of unusedNames) {
             console.log(`❌${unusedName}`)
